+++ src/main/java/nextstep/dto/MoveResult.java
@@ -1,4 +1,4 @@
-package nextstep.domain;
+package nextstep.dto;
 
 import java.util.Objects;
