+++ src/test/java/nextstep/domain/CarTest.java
@@ -1,5 +1,6 @@
 package nextstep.domain;
 
+import nextstep.dto.MoveResult;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.ValueSource;
