+++ src/test/java/nextstep/domain/CarsTest.java
@@ -1,5 +1,6 @@
 package nextstep.domain;
 
+import nextstep.dto.MoveResult;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
