+++ src/test/java/nextstep/domain/MoveResultTest.java
@@ -1,5 +1,6 @@
 package nextstep.domain;
 
+import nextstep.dto.MoveResult;
 import org.junit.jupiter.api.Test;
 
 import static org.assertj.core.api.Assertions.assertThat;
